+++ autofactory/django/builders/concrete.py
@@ -5,6 +5,7 @@
 # Distributed under MIT License. See LICENSE file for details.
 from __future__ import unicode_literals
 
+import factory
 from factory import fuzzy
 
 
@@ -13,4 +14,7 @@ def from_choices(field_cls):
 
 
 def from_default(field_cls):
+    if callable(field_cls.default):
+        return factory.LazyFunction(field_cls.default)
+
     return field_cls.default