+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 - Change generated with `autofactory` name from `GeneratedModelFactory` to 
 `ModelFactory`.
+- Fix callable as a `django.db.models.Field.default` not being called on 
+instance generation.
 
 ## [0.2.0] — 2019-02-08
 - Add builders registry for Django ORM.